f64 -> f32

+++ syzygy/src/hittable.rs
@@ -5,13 +5,13 @@ use crate::material::Material;
 pub struct Hit<'a> {
     pub point: Vec3,
     pub normal: Vec3,
-    pub t: f64,
+    pub t: f32,
     pub front_face: bool,
     pub mat: &'a dyn Material,
 }
 
 impl<'a> Hit<'a> {
-    pub fn new(point: Vec3, normal: Vec3, t: f64, front_face: bool, mat: &'a dyn Material) -> Hit<'a> {
+    pub fn new(point: Vec3, normal: Vec3, t: f32, front_face: bool, mat: &'a dyn Material) -> Hit<'a> {
         Hit {
             point, 
             normal,
@@ -23,17 +23,17 @@ impl<'a> Hit<'a> {
 }
 
 pub trait Hittable {
-    fn intersect(&self, r: &Ray, tmin: f64, tmax: f64) -> Option<Hit>;
+    fn intersect(&self, r: &Ray, tmin: f32, tmax: f32) -> Option<Hit>;
 }
 
 pub struct Sphere<M: Material> {
     center: Vec3,
-    radius: f64,
+    radius: f32,
     mat: M,
 }
 
 impl<M: Material> Sphere<M> {
-    pub fn new(center: Vec3, radius: f64, mat: M) -> Sphere<M> {
+    pub fn new(center: Vec3, radius: f32, mat: M) -> Sphere<M> {
         Sphere {
             center,
             radius,
@@ -44,7 +44,7 @@ impl<M: Material> Sphere<M> {
 
 impl<M: Material> Hittable for Sphere<M> {
 
-    fn intersect(&self, r: &Ray, tmin: f64, tmax: f64) -> Option<Hit> {
+    fn intersect(&self, r: &Ray, tmin: f32, tmax: f32) -> Option<Hit> {
         let oc = r.o - self.center;
         let a = r.d.dot(&r.d);
         let half_b = oc.dot(&r.d);
@@ -83,7 +83,7 @@ impl Hittables {
 }
 
 impl Hittable for Hittables {
-    fn intersect(&self, r: &Ray, tmin: f64, tmax: f64) -> Option<Hit> {
+    fn intersect(&self, r: &Ray, tmin: f32, tmax: f32) -> Option<Hit> {
         self.0.iter().fold(None, |acc, curr| 
             curr.intersect(r, tmin, acc.as_ref().map_or(tmax, |x| x.t)).or(acc)
         )
